Update to use SSB-Config by default
Problem: Having to use your own SSB-Config is a bit silly, since we're
all doing it.

Solution: Make a breaking change and use SSB-Config by default. While
we're at it, let's add a readme and Common-Good and a quick sanity test
to make sure the server can at least start and stop.